nix: fix Julia PyCall not linking to environment python
PyCall relies on hardcoded values stored in deps/deps.jl which tell
where python and libpython are found. These hardcoded values are set
when building PyCall.

This is a non-issue on most systems, but due to the nature of nix, each
change of our python environment moves it to another location, breaking
PyCall.

This introduces a workaround to generate the deps.jl from nix and
replace the one in the local julia depo with a symlink to the generated
one, ensuring that rebuilds of the python environment do not break the
development shell.