nix: make RCall.jl work with R
nix's rWrapper uses R_LIBS_SITE to set the location of all libraries
added. This R_LIBS_SITE is not accessible outside the value set in the
rWrapper, so we have to run R to get the environment value from the
wrapper and export it so that libR.so used by julia's RCall finds the
libraries.

Maybe we should reconsider how the R_LIBS_SITE is set up so that a more
sensible method is possible.